Refactor email handling: update import paths, add OriginalEmail schema, and remove unused parsing utility
